Update documentation and move code to its own folder
