+++ audio/Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+watchdog = "==2.1.6"
+pyyaml = "==6.0"
+
+[packages]
+pygame = "==2.1.2"
+websocket-client = "==1.2.3"
+
+[requires]
+python_version = "3.8"
